+++ Stride.CommunityToolkit.Bepu/Engine/Primitive2DCreationOptionsWithBepu.cs
@@ -1,9 +1,9 @@
-using Stride.BepuPhysics;
 using Stride.BepuPhysics.Definitions.Colliders;
-using Stride.CommunityToolkit.Bepu;
+using Stride.BepuPhysics;
+using Stride.CommunityToolkit.Engine;
+using Stride.Core.Mathematics;
 
-namespace Stride.CommunityToolkit.Engine;
-
+namespace Stride.CommunityToolkit.Bepu.Engine;
 public class Primitive2DCreationOptionsWithBepu : PrimitiveCreationOptions
 {
     /// <summary>
