+++ Stride.CommunityToolkit.Bepu/Stride.CommunityToolkit.Bepu.csproj
@@ -1,0 +1,17 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net8.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Stride.BepuPhysics" Version="0.9.2" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\src\Stride.CommunityToolkit\Stride.CommunityToolkit.csproj" />
+  </ItemGroup>
+
+</Project>
