+++ examples/code-only/Example_Bepu_Playground/Example_Bepu_Playground.csproj
@@ -14,6 +14,7 @@
     <ItemGroup>
         <ProjectReference Include="..\..\..\src\Stride.CommunityToolkit.Skyboxes\Stride.CommunityToolkit.Skyboxes.csproj" />
         <ProjectReference Include="..\..\..\src\Stride.CommunityToolkit.Windows\Stride.CommunityToolkit.Windows.csproj" />
+        <ProjectReference Include="..\..\..\Stride.CommunityToolkit.Bepu\Stride.CommunityToolkit.Bepu.csproj" />
         <ProjectReference Include="..\Example.Common\Example.Common.csproj" />
     </ItemGroup>
