+++ src/Stride.CommunityToolkit/Stride.CommunityToolkit.csproj
@@ -18,7 +18,6 @@
         <PackageReference Include="Stride.Particles" Version="4.2.0.2232" />
         <PackageReference Include="Stride.Physics" Version="4.2.0.2232" />
         <PackageReference Include="Stride.UI" Version="4.2.0.2232" />
-        <PackageReference Include="Stride.BepuPhysics" Version="0.9.2" />
     </ItemGroup>
 
     <ItemGroup>
